core: compute max speed envelope in "buildFinalEnvelope"
